Update get_weather.py
Added some code to make it prettify the XLSX contents. Now it not only auto-widths the rows, but centers the data.

Also added a warning about the data source. People outside of Japan should find a different source outside of Japan, but this can be found at the same website listed in the file.