Add files via upload
This is an updated version of the program which now adds AQI data to the output XLSX file.
Version 2 now pulls down PM 2.5, PM 10.0, CO, NO2, SO2, Ozone data and calculates the AQI based on these values.

As before, the program grabs data for my location in Shizuoka, Japan. You will need to update the latitutude and longitude information in the program to match your location.